requirements.txt
# ──────────────────────────────────────────────
# Python Dependencies
# SteamFullMarket Analytics Pipeline
# ──────────────────────────────────────────────
# Airflow
apache-airflow==2.9.1
apache-airflow-providers-google==10.21.0
# Database
psycopg2-binary==2.9.9
# Data processing
pandas==2.2.2
pyarrow==16.0.0
# Data ingestion
kaggle==2.0.1
# Google Cloud
google-cloud-storage==2.16.0
google-cloud-bigquery==3.21.0
# dbt
dbt-bigquery==1.11.8
# Utilities
python-dotenv==1.0.1